Added local database simulation

+++ lib/feature_login/login_page.dart
@@ -1,7 +1,9 @@
+import 'dart:io';
+
+import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
-import 'package:isar/isar.dart';
 import 'package:noteit/core/constants/constants.dart';
-import 'package:noteit/entities/credentials.dart';
+import 'package:noteit/entities/user.dart';
 
 import '../main.dart';
 
@@ -16,12 +18,28 @@ class _LoginPageState extends State<LoginPage> {
   int selectedAccountIndex = -1; // Default to the first account
   String currUsername = '';
   String currPassword = '';
+  String currProfilePicture = '';
+
+  void _changeProfilePicture() async {
+    FilePickerResult? result = await FilePicker.platform.pickFiles(
+      type: FileType.image,
+      allowMultiple: false,
+    );
+
+    if (result != null) {
+      File file = File(result.files.single.path!);
+
+      setState(() {
+        currProfilePicture = file.path;
+      });
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text('Login'),
+        title: const Text('Login'),
       ),
       body: Center(
         child: Padding(
@@ -34,7 +52,7 @@ class _LoginPageState extends State<LoginPage> {
                 mainAxisAlignment: MainAxisAlignment.center,
                 children: [
                   // Add a delete button (only if an account is selected)
-                  if (selectedAccountIndex != -1)
+                  if (selectedAccountIndex != -1 && selectedAccountIndex != -2)
                     IconButton(
                       onPressed: () async {
                         await isarService.deleteCredentials(
@@ -43,7 +61,7 @@ class _LoginPageState extends State<LoginPage> {
                           selectedAccountIndex = -1;
                         });
                       },
-                      icon: Icon(Icons.delete),
+                      icon: const Icon(Icons.delete),
                     ),
                   SizedBox(width: MediaQuery.of(context).size.width * 0.05),
                   FutureBuilder<List<DropdownMenuItem<int>>>(
@@ -51,7 +69,7 @@ class _LoginPageState extends State<LoginPage> {
                     builder: (context, snapshot) {
                       if (snapshot.connectionState == ConnectionState.waiting) {
                         // Loading indicator or placeholder
-                        return CircularProgressIndicator();
+                        return const CircularProgressIndicator();
                       } else if (snapshot.hasError) {
                         // Error handling
                         return Text('Error: ${snapshot.error}');
@@ -71,17 +89,53 @@ class _LoginPageState extends State<LoginPage> {
                   ),
                 ],
               ),
-              SizedBox(height: 16),
+              const SizedBox(height: 16),
               if (selectedAccountIndex == -1) ...[
                 TextField(
                   controller: _usernameController,
-                  decoration: InputDecoration(
+                  decoration: const InputDecoration(
                     labelText: 'Username',
                   ),
                 ),
                 TextField(
                   controller: _passwordController,
-                  decoration: InputDecoration(
+                  decoration: const InputDecoration(
+                    labelText: 'Password',
+                  ),
+                ),
+              ],
+              if (selectedAccountIndex == -2) ...[
+                InkWell(
+                  onTap: _changeProfilePicture,
+                  child: Column(
+                    mainAxisAlignment: MainAxisAlignment.center,
+                    crossAxisAlignment: CrossAxisAlignment.center,
+                    children: [
+                      if (currProfilePicture.isNotEmpty)
+                        CircleAvatar(
+                          radius: 50,
+                          backgroundImage: FileImage(File(currProfilePicture)),
+                        )
+                      else
+                        const CircleAvatar(
+                          radius: 50,
+                          child: Icon(Icons.person),
+                        ),
+                      SizedBox(
+                        height: MediaQuery.of(context).size.height * 0.04,
+                      ),
+                    ],
+                  ),
+                ),
+                TextField(
+                  controller: _usernameController,
+                  decoration: const InputDecoration(
+                    labelText: 'Username',
+                  ),
+                ),
+                TextField(
+                  controller: _passwordController,
+                  decoration: const InputDecoration(
                     labelText: 'Password',
                   ),
                 ),
@@ -89,52 +143,95 @@ class _LoginPageState extends State<LoginPage> {
               SizedBox(height: MediaQuery.of(context).size.height * 0.05),
               ElevatedButton(
                 onPressed: () async {
-                  if (selectedAccountIndex == -1) {
-                    username = _usernameController.text;
-                    password = _passwordController.text;
-                    // Check if username or password is empty
-                    if (username.isEmpty || password.isEmpty) {
-                      // Show an alert dialog
-                      showDialog(
-                        context: context,
-                        builder: (context) => AlertDialog(
-                          title: Text('Error'),
-                          content:
-                              Text('Username or password cannot be empty.'),
-                          actions: [
-                            TextButton(
-                              onPressed: () {
-                                Navigator.pop(context);
-                              },
-                              child: Text('OK'),
-                            ),
-                          ],
-                        ),
-                      );
-                      return;
-                    }
+                  switch (selectedAccountIndex) {
+                    case -1:
+                      username = _usernameController.text;
+                      password = _passwordController.text;
+                      // Check if username or password in database
+                      if (await isarService.verifyCredentials(
+                          username, password)) {
+                        // Login
+                        // ignore: use_build_context_synchronously
+                        Navigator.pushReplacement(
+                          context,
+                          MaterialPageRoute(builder: (context) => AppState()),
+                        );
+                      } else {
+                        // Show an alert dialog
+                        // ignore: use_build_context_synchronously
+                        showDialog(
+                          context: context,
+                          builder: (context) => AlertDialog(
+                            title: const Text('Error'),
+                            content: const Text(
+                                'Username or password is incorrect.'),
+                            actions: [
+                              TextButton(
+                                onPressed: () {
+                                  Navigator.pop(context);
+                                },
+                                child: const Text('OK'),
+                              ),
+                            ],
+                          ),
+                        );
+                      }
+                      break;
+                    case -2:
+                      username = _usernameController.text;
+                      password = _passwordController.text;
+                      profilePicture = currProfilePicture;
+                      // Check if username or password is empty
+                      if (username.isEmpty ||
+                          password.isEmpty ||
+                          profilePicture.isEmpty) {
+                        // Show an alert dialog
+                        showDialog(
+                          context: context,
+                          builder: (context) => AlertDialog(
+                            title: const Text('Error'),
+                            content: const Text(
+                                'Username or password cannot be empty or profile picture is not selected.'),
+                            actions: [
+                              TextButton(
+                                onPressed: () {
+                                  Navigator.pop(context);
+                                },
+                                child: const Text('OK'),
+                              ),
+                            ],
+                          ),
+                        );
+                        return;
+                      }
 
-                    Credentials credentials = Credentials(username, password);
-                    await isarService.saveCredentials(credentials);
-                    Navigator.pushReplacement(
-                      context,
-                      MaterialPageRoute(builder: (context) => AppState()),
-                    );
-                  } else {
-                    // Login
-                    Credentials? credentials = await isarService.getCredentials(
-                        credentialsList[selectedAccountIndex].id);
-                    if (credentials != null) {
-                      username = credentials.username;
-                      password = credentials.password;
+                      UserEntity credentials =
+                          UserEntity(username, password, profilePicture);
+                      await isarService.saveCredentials(credentials);
+                      // ignore: use_build_context_synchronously
                       Navigator.pushReplacement(
                         context,
                         MaterialPageRoute(builder: (context) => AppState()),
                       );
-                    }
+                      break;
+                    default:
+                      // Login
+                      UserEntity? credentials =
+                          await isarService.getCredentials(
+                              credentialsList[selectedAccountIndex].id);
+                      if (credentials != null) {
+                        username = credentials.username;
+                        password = credentials.password;
+                        profilePicture = credentials.profilePicture;
+                        Navigator.pushReplacement(
+                          context,
+                          MaterialPageRoute(builder: (context) => AppState()),
+                        );
+                      }
+                      break;
                   }
                 },
-                child: Text('Login'),
+                child: const Text('Login'),
               ),
             ],
           ),
@@ -148,9 +245,16 @@ class _LoginPageState extends State<LoginPage> {
     credentialsList = await isarService.getAllCredentials();
 
     items.add(
-      DropdownMenuItem<int>(
+      const DropdownMenuItem<int>(
         value: -1,
-        child: Text('Add account'),
+        child: Text('Login'),
+      ),
+    );
+
+    items.add(
+      const DropdownMenuItem<int>(
+        value: -2,
+        child: Text('Create account'),
       ),
     );
 
@@ -158,7 +262,17 @@ class _LoginPageState extends State<LoginPage> {
       items.add(
         DropdownMenuItem<int>(
           value: i,
-          child: Text('${credentialsList[i].username}'),
+          child: Row(
+            children: [
+              if (credentialsList[i].profilePicture.isNotEmpty)
+                CircleAvatar(
+                  backgroundImage:
+                      FileImage(File(credentialsList[i].profilePicture)),
+                ),
+              const SizedBox(width: 8),
+              Text('${credentialsList[i].username}'),
+            ],
+          ),
         ),
       );
     }
